Began implementing ics by clicking.
Now can click on axis in IC tab and start changing an array of substrate values. The value is only ever changing from 0 to 1. This array is then saved in a CSV as is, which is ultimately not helpful, but getting it into the proper format shouldn't be hard. Just need to integrate with the Microenvironment tab