Rectangles!!!
These work super well, not as susceptible to fast moving cursors missing voxels. They go up to, but not over the edges.

If the voxels are dx apart in the x direction, then 0<=R1<dx means is a one-voxel-wide rectangle. If dx<=R1<2dx, then it is a 3-voxel-wide rectangle. Etc. Similarly for R2 and the y-direction.