Gaussians for the win!
These are so cool! Create a Gaussian distribution with a given sigma and a rectangle to bound it on (for computational efficienty, I guess? I originally intended it for finer control, but now I'm not sure it's necessary). Rather than simply override, it takes the max value so that you can draw paths and now have points "washed out" in the wake of your cursor as the lower values of the Gaussian hit them.