Save nonzero substrate ics with header row
Only saving those substrate ics for which there is some nonzero value in the grid. To specify which one(s), including a header row that must start x,y,z and then be the nonzero substrates in the order they appear in the substrate_list. PhysiCell will read this header row to assign these value to the proper substrate.

If all substrate values are currently 0, then do not bother saving anything, just disable these ics.