Auto add pd dmg var to custom data, I think?
I honestly cannot even recall what the bug was. Writing this commit literally after finishing, and have no idea...

Or maybe this was all about automatically adding S_damage to custom_data tab. Yeah...I'm going with that.

Turns out studio has a master list of all custom data and it just always has "" as a key. So ignore this somehow.

Anyways, now whenever a PD model (not None) is selected for a cell, a damage variable is added to the custom_data tab, provided it's not already there. If a PD model is set to None, studio checks if any cell type considers the in-focus substrate as a PD substrate. If not, the damage variable is removed.

Users can include the PD stuff without tying to a rule. It won't make a difference unless they do more.

I'm not sure what happens if the damage variable is accidentally deleted (misclick and suddenly no cells have S as a PD substrate) and rules were already written regarding S_damage. I'll check on that now.
And I checked and rules does not change. That's good. It's not overly sensitive.