I got it! Wedge check implemented!!
Break it up into cases: center inside domain or outside.
If inside, then everything along ray from center to boundary of domain is in the domain, so just see if one of the rays to the boundary is further than the inner radius.
The only points worth checking are those at angles theta1, theta2 and any rays to corners found between theta1 and theta2. I check cardinal directions as well, just because those are so fast anyways.

If outside, first reorient (reflect or rotate) so that the center is either left of the domain or bottom-left of domain.
In either case, collect the rays that pass through key points of the domain: theta1, theta2, and through corners and directly right (in case where center is directly left, not down-and-left).
Consider a ray that moves counter-clockwise from one of these key points to the next.
Along that ray, should it intersect the domain (only a question for theta1 and theta2), draw the interval on the ray corresponding to the distances along the ray that intersect the domain.
The endpoints of this interval must move monotonically between these key points (not necessarily each in the same direction, but left goes one way the whole time, same for right)
So, these points are sufficient for checking if at any point the wedge intersects the domain, for the following reason.
If wedge overlaps with the domain along a ray if and only if the interval described above overlaps with the interval (r0,r1) defined by the wedge.
If one of these points has such an overlap, then we know the wedge is partially inside the domain.
Additionally, if the above interval switches from entirely above (r0,r1) to below between consecutive key points, then it must intersect the domain.
The same goes for if the switch is from entirely below to entirely above.
Finally, caution must be taken that if the ray sweeping counter-clockwise passes the corner of the top-left corner of the domain and thus exits the domain, that this check of switching is disabled until after re-entering the domain at the first corner we hit (which depends on whether we are left or bottom-left of the domain), should that occur.

That hopefully conveys the logic that this code attempts to put into practice. Hopefully someone reads this and finds it useful some day. Even me, I'll accept that as a win. Or maybe I should copy-paste this somewhere? Right, this is a commit message on 2024-02-21 at 11:36AM ET. Now I'm just ramb