Better sampling for round regions
When clicking "Plot" button, shrink out radius if possible and stretch inner radius (if applicable) so that sampling is done more concisely.

Also, found out that matplotlib.patches.Annulus has a quirk where the width cannot match the semi-minor axis. In other words, it forces a hole in the middle. The error message didn't comport with this. PR submitted to them to fix the check (not the message) so that annuli can have a "filled in" hole. I checked that it can close the hole and re-grow the hole if that inner radius is changed. So, it shouldn't be a problem on their end, for that at least.