Create WarningWindow class and check for name conflicts
If the names conflict, a warning appears listing the names that are identical. I'm still not sure what happens if the user goes forward with repeat names, so maybe this will eventaully just force the user back anyways? At least now the warning is hereCreate WarningWindow class and check for name conflicts

If the names conflict, a warning appears listing the names that are identical. I'm still not sure what happens if the user goes forward with repeat names, so maybe this will eventaully just force the user back anyways? At least now the warning is here.